feat: Use AG chart histogram.
This reverts commit 92651015d52c02532d32819bd14e9548b4b1a981.
